Use correct EC command to fetch keyboard backlight level
The EC has a dedicated command for fetching the backlight level; use
that instead of checking PWM duty cycle, which doesn't appear to work
(catting the 'brightness' sysfs node always returns 0, regardless of the
actual backlight level).

This is also what framework_tool does when passed '--kblight' with no
options.